+++ src/controllers/artistController.ts
@@ -3,6 +3,10 @@ import ArtistService from '../services/ArtistService';
 import { ArtistEntity } from '../repository/database/entities/artist';
 import BaseController from './baseController';
 import { AlbumEntity } from '../repository/database/entities/album';
+import ArtistValidator from './validators/artistValidator';
+import httpException from './../common/httpException';
+import AlbumValidator from './validators/albumValidator';
+import Constants from '../config/constants/Constants';
 
 export default class ArtistController extends BaseController {
   private readonly _artistService: ArtistService
@@ -13,6 +17,9 @@ export default class ArtistController extends BaseController {
 
   create = (req: Request, res: Response, next: NextFunction): void => {
     const artist: ArtistEntity = req.body as ArtistEntity;
+    if (!ArtistValidator.validate(artist))
+      throw new httpException(500, Constants.INVALID_PARAMETERS_ERROR);
+
     this.handleExceptionAndExecuteRequest(() => {
       this._artistService.create(artist, (error: Error, result: any) => this.sendResponse(res, error, result, next));
     });
@@ -21,6 +28,9 @@ export default class ArtistController extends BaseController {
   addAlbum = (req: Request, res: Response, next: NextFunction): void => {
     const artistId: string = req.params.id as string;
     const album: AlbumEntity = req.body as AlbumEntity;
+
+    if (!AlbumValidator.validate(album) || !artistId)
+      throw new httpException(500, Constants.INVALID_PARAMETERS_ERROR);
 
     this.handleExceptionAndExecuteRequest(() => {
       this._artistService.addAlbum(artistId, album, (error: Error, result: any) => this.sendResponse(res, error, result, next));
@@ -31,6 +41,9 @@ export default class ArtistController extends BaseController {
     const albumId: string = req.params.albumId as string;
     const artistId: string = req.params.artistId as string;
 
+    if (!artistId || !albumId)
+      throw new httpException(500, Constants.INVALID_PARAMETERS_ERROR);
+
     this.handleExceptionAndExecuteRequest(() => {
       this._artistService.removeAlbum(artistId, albumId, (error: Error, result: any) => this.sendResponse(res, error, result, next));
     });
@@ -38,6 +51,8 @@ export default class ArtistController extends BaseController {
 
   delete = (req: Request, res: Response, next: NextFunction): void => {
     const artistId: string = req.params.id as string;
+    if (!artistId)
+      throw new httpException(500, Constants.INVALID_PARAMETERS_ERROR);
 
     this.handleExceptionAndExecuteRequest(() => {
       this._artistService.delete(artistId, (error: Error, result: any) => this.sendResponse(res, error, result, next));
@@ -48,6 +63,9 @@ export default class ArtistController extends BaseController {
   update = (req: Request, res: Response, next: NextFunction): void => {
     const artistModel: ArtistEntity = req.body as ArtistEntity;
     const artistId: string = req.params.id as string;
+
+    if (!ArtistValidator.validate(artistModel) || !artistId)
+      throw new httpException(500, Constants.INVALID_PARAMETERS_ERROR);
 
     this.handleExceptionAndExecuteRequest(() => {
       this._artistService.update(artistId, artistModel, (error: Error, result: any) => this.sendResponse(res, error, result, next));
@@ -61,9 +79,12 @@ export default class ArtistController extends BaseController {
   }
 
   getById = (req: Request, res: Response, next: NextFunction): void => {
-    const albumId: string = req.params.id as string;
+    const artistId: string = req.params.id as string;
+    if (!artistId)
+      throw new httpException(500, Constants.INVALID_PARAMETERS_ERROR);
+
     this.handleExceptionAndExecuteRequest(() => {
-      this._artistService.findById(albumId, (error: Error, result: any) => this.sendResponse(res, error, result, next));
+      this._artistService.findById(artistId, (error: Error, result: any) => this.sendResponse(res, error, result, next));
     });
   }
 }
